+++ tests/unit/grammar/GrammarTest.php
@@ -10,18 +10,24 @@ class GrammarTest extends BaseTestCase
     public function testCreateFailManyMainRules()
     {
         $this->expectException(GrammarException::class);
-        TextLoader::createGrammar("E: A \$; A: a; E: B \$; B: b");
+        TextLoader::createGrammar('E: A $; A: a; E: B $; B: b');
     }
 
     public function testCreateFailNoMainRule()
     {
         $this->expectException(GrammarException::class);
-        TextLoader::createGrammar("A: A a; A: a");
+        TextLoader::createGrammar('A: A a; A: a');
     }
 
     public function testCreateFailNoTerminals()
     {
         $this->expectException(GrammarException::class);
-        TextLoader::createGrammar("E: A \$; A: B; B: A");
+        TextLoader::createGrammar('E: A $; A: B; B: A');
+    }
+
+    public function testCreateFailUndefined()
+    {
+        $this->expectException(GrammarException::class);
+        TextLoader::createGrammar('E: A $; A: foo; A: bar');
     }
 }