+++ tests/unit/grammar/exporter/JsonExporterTest.php
@@ -24,12 +24,12 @@ class JsonExporterTest extends BaseTestCase
     {
         return [
             [
-                TextLoader::createGrammar('G: a $'),
-                '{"rules":[{"name":"G","eof":true,"definition":["a"]}],"terminals":[{"name":"a"}]}'
+                TextLoader::createGrammar('G: a $; a: /a+/'),
+                '{"rules":[{"name":"G","eof":true,"definition":["a"]}],"terminals":[{"name":"a","match":"a+"}]}'
             ],
             [
-                TextLoader::createGrammar('G: A $ ; A (loop) : A a ; A: a; A: b; A: "x"; A: .c; a: "y"; b: /\\d+/'),
-                '{"rules":[{"name":"G","eof":true,"definition":["A"]},{"name":"A","tag":"loop","definition":["A","a"]},{"name":"A","definition":["a"]},{"name":"A","definition":["b"]},{"name":"A","definition":["x"]},{"name":"A","definition":[{"name":"c","hidden":true}]}],"terminals":[{"name":"a","match":"y","isText":true},{"name":"b","match":"\\\\d+"},{"name":"c"},"x"]}'
+                TextLoader::createGrammar('G: A $ ; A (loop) : A a ; A: a; A: b; A: "x"; A: .c; a: "y"; b: /\\d+/; c: /c+/'),
+                '{"rules":[{"name":"G","eof":true,"definition":["A"]},{"name":"A","tag":"loop","definition":["A","a"]},{"name":"A","definition":["a"]},{"name":"A","definition":["b"]},{"name":"A","definition":["x"]},{"name":"A","definition":[{"name":"c","hidden":true}]}],"terminals":[{"name":"a","match":"y","isText":true},{"name":"b","match":"\\\\d+"},{"name":"c","match":"c+"},"x"]}'
             ],
         ];
     }